backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
boto3==1.25.4
botocore==1.28.4
CacheControl==0.12.11
cached-property==1.5.2
cachetools==5.2.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
cryptography==38.0.3
Django==3.2.16
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-sslserver==0.22
django-storages==1.13.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
ffmpeg==1.4
firebase-admin==6.0.1
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
importlib-metadata==4.13.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
MarkupSafe==2.1.1
msgpack==1.0.4
packaging==21.3
pilkit==2.0
Pillow==9.1.1
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
zipp==3.10.0
